requirements.txt
python-dotenv==1.1.0
teradatasql==20.0.0.32
